FaceTrack Attendance System/requirements.txt
flask==2.3.3
opencv-python==4.7.0.72
face-recognition==1.3.0
numpy==1.24.3
pandas==2.0.1
openpyxl==3.1.2
Pillow==9.5.0
tkinter
matplotlib==3.7.1
streamlit==1.25.0
plotly==5.15.0
bcrypt==4.0.1